Implemented the checkMove method in the parent class CheckPlayerMove rather than in its subclasses. Created class checkKingMove.
