+++ src/UseCases/CheckBishopMove.java
@@ -1,6 +1,5 @@
 package UseCases; 
 
-import Entities.Bishop;
 import Entities.ChessPiece; 
 import Other.GameState; 
 
@@ -71,19 +70,6 @@ public class CheckBishopMove extends CheckPlayerMove {
 
         int[][] array = result.toArray();
         return array;
-    }
-
-    public boolean checkMove(int newRow, int newColumn, ChessPiece bishop, GameState gameState) {
-        int[] desiredMove = {newRow, newColumn};
-        int[][] possibleMoves = validMoves((Bishop) bishop, gameState);
-        
-        for (int[] move : possibleMoves) {
-            if (Arrays.equals(desiredMove, move)) {
-                return true;
-            }
-        }
-
-        return false;
     }
   
 }
