+++ src/UseCases/CheckPawnMove.java
@@ -1,6 +1,5 @@
 package UseCases;
 
-import Entities.Pawn;
 import Entities.ChessPiece;
 import Other.GameState;
 
@@ -66,16 +65,4 @@ public class CheckPawnMove extends CheckPlayerMove {
         return result;
     }
      
-    public boolean checkMove(int newRow, int newColumn, ChessPiece pawn, GameState gameState) {
-        int[] desiredMove = {newRow, newColumn};
-        int[][] possibleMoves = validMoves((Pawn) pawn, gameState);
-        
-        for (int[] move : possibleMoves) {
-            if (Arrays.equals(desiredMove, move)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
 }
