+++ src/UseCases/CheckRookMove.java
@@ -1,6 +1,5 @@
 package UseCases; 
 
-import Entities.Rook;
 import Entities.ChessPiece; 
 import Other.GameState; 
 
@@ -64,19 +63,6 @@ public class CheckRookMove extends CheckPlayerMove {
         int[][] array = result.toArray(); 
         return array; 
 
-    }
-
-    public boolean checkMove(int newRow, int newColumn, ChessPiece rook, GameState gameState) {
-        int[] desiredMove = {newRow, newColumn};
-        int[][] possibleMoves = validMoves((Rook) rook, gameState);
-        
-        for (int[] move : possibleMoves) {
-            if (Arrays.equals(desiredMove, move)) {
-                return true;
-            }
-        }
-
-        return false;
     }
   
 }
